requirements.txt
## The following requirements were added by pip freeze:
astroid==1.4.5
click==6.6
colorama==0.3.7
docopt==0.6.2
Flask==0.11.1
Flask-Injector==0.7.1
Flask-Login==0.3.2
Flask-OAuthlib==0.9.3
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
furl==0.5.1
injector==0.9.1
itsdangerous==0.24
Jinja2==2.8
lazy-object-proxy==1.2.2
MarkupSafe==0.23
oauthlib==1.1.2
orderedmultidict==0.7.6
passlib==1.6.5
pep8==1.7.0
py==1.4.31
pycrypto==2.6.1
pylint==1.5.5
pytest==2.9.2
PyYAML==3.11
requests==2.10.0
requests-oauthlib==0.6.1
simple-crypt==4.1.7
six==1.10.0
SQLAlchemy==1.0.13
SQLAlchemy-Utils==0.32.9
uWSGI==2.0.13.1
Werkzeug==0.11.10
wrapt==1.10.8
WTForms==2.1
